+++ src/constants.ts
@@ -1,5 +1,5 @@
 export const EXTENSION_NAME = "codetour";
 export const FS_SCHEME = EXTENSION_NAME;
 export const ICON_URL =
-  "https://cdn.jsdelivr.net/gh/vsls-contrib/code-tour/images/icon.png";
+  "https://cdn.jsdelivr.net/gh/vsls-contrib/code-tour/images/icon-small.png";
 export const VSCODE_DIRECTORY = ".vscode";