+++ src/extension.ts
@@ -13,6 +13,7 @@ import { discoverTours } from "./store/provider";
 import { registerTreeProvider } from "./tree";
 import { registerDecorators } from "./decorator";
 import { store } from "./store";
+import { registerNotebookProvider } from "./notebook";
 
 export async function activate(context: vscode.ExtensionContext) {
   registerCommands();
@@ -46,6 +47,7 @@ export async function activate(context: vscode.ExtensionContext) {
   registerTreeProvider(context.extensionPath);
   registerFileSystemProvider();
   registerStatusBar();
+  registerNotebookProvider();
 
   return {
     startTour: startCodeTour,